pyproject.toml
[project]
name = "nexus-pdf"
version = "0.1.0"
description = "A modular PDF processing service that converts PDFs to images and analyzes them with AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.116.1",
    "openai>=1.97.1",
    "pdf2image>=1.17.0",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.1",
    "rq>=2.4.1",
    "uvicorn>=0.35.0",
    "redis>=5.0.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    "Pillow>=10.0.0",
    "poppler-utils>=23.0.0",
]